+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 IMPROVEMENTS:
 
 * Use Go 1.20
+* Replace archived FreeIPA Go library with new maintained fork
 
 ## 0.7.0 (October 29, 2020)